EntityCrossProjectSelectionConfirmModal: move to @labkey/components/entities
